Add docs for the low-level/fundamental types. Did a good bit of code re-org, too, cleaning out cruft.
