+++ configure.ac
@@ -35,11 +35,15 @@ AX_CXX_COMPILE_STDCXX([17],[noext],[mandatory])
 fi
 
 AC_LANG(C++)
-AC_CONFIG_FILES([Makefile fwdpp/version.hpp fwdpp/Makefile fwdpp/io/Makefile fwdpp/algorithm/Makefile fwdpp/io/detail/Makefile
-                 fwdpp/internal/Makefile fwdpp/tags/Makefile fwdpp/sugar/Makefile fwdpp/poptypes/Makefile
-                 fwdpp/gsl/Makefile fwdpp/extensions/Makefile fwdpp/meta/Makefile
-				 fwdpp/simfunctions/Makefile
-                 fwdpp/ts/Makefile fwdpp/ts/detail/Makefile fwdpp/ts/marginal_tree_functions/Makefile
+AC_CONFIG_FILES([Makefile fwdpp/version.hpp fwdpp/Makefile
+                 fwdpp/fundamental_types/Makefile fwdpp/io/Makefile
+                 fwdpp/algorithm/Makefile fwdpp/io/detail/Makefile
+                 fwdpp/internal/Makefile fwdpp/tags/Makefile
+                 fwdpp/sugar/Makefile fwdpp/poptypes/Makefile
+                 fwdpp/gsl/Makefile fwdpp/extensions/Makefile
+                 fwdpp/meta/Makefile fwdpp/simfunctions/Makefile
+                 fwdpp/ts/Makefile fwdpp/ts/detail/Makefile
+                 fwdpp/ts/marginal_tree_functions/Makefile
                  fwdpp/ts/simplification/Makefile
                  fwdpp/ts/recording/Makefile
 				 fwdpp/util/Makefile
