+++ doc/rst/fundamentaltypes.rst
@@ -1,0 +1,6 @@
+Introduction
+=====================================
+
+This section defines the lowest-level types required by a simulation.
+
+These types can be included all at once via :ref:`file_fwdpp_forward_types.hpp`.
