+++ doc/rst/typedefs.rst
@@ -1,0 +1,8 @@
+.. _typedefs:
+
+Typedefs
+=====================================
+
+The following ``typedefs`` are used throughout ``fwdpp``:
+
+* :type:`fwdpp::uint_t` represents an unsigned integer type.
