+++ fwdpp/Makefile.am
@@ -1,6 +1,6 @@
 pkgincludedir=$(prefix)/include/fwdpp
 
-SUBDIRS=internal tags sugar gsl extensions io meta algorithm ts poptypes simfunctions genetic_map util
+SUBDIRS=internal fundamental_types tags sugar gsl extensions io meta algorithm ts poptypes simfunctions genetic_map util
 pkginclude_HEADERS=debug.hpp \
 	diploid.hh \
 	fitness_models.hpp \
