+++ fwdpp/forward_types_serialization.hpp
@@ -8,34 +8,6 @@ namespace fwdpp
 {
     namespace io
     {
-        template <> struct serialize_mutation<mutation>
-        {
-            template <typename streamtype>
-            inline void
-            operator()(streamtype &buffer, const mutation &m) const
-            {
-                io::scalar_writer writer;
-                writer(buffer, &m.pos);
-                writer(buffer, &m.s);
-                writer(buffer, &m.h);
-            }
-        };
-
-        template <> struct deserialize_mutation<mutation>
-        {
-            template <typename streamtype>
-            inline mutation
-            operator()(streamtype &buffer) const
-            {
-                double pos, s, h;
-                io::scalar_reader reader;
-                reader(buffer, &pos);
-                reader(buffer, &s);
-                reader(buffer, &h);
-
-                return mutation(pos, s, h);
-            }
-        };
         template <> struct serialize_haploid_genome<haploid_genome>
         /// \brief Serialize a haploid_genome
         ///
@@ -68,7 +40,7 @@ namespace fwdpp
         {
             template <typename streamtype>
             inline haploid_genome
-            operator()(streamtype& buffer) const
+            operator()(streamtype &buffer) const
             {
                 scalar_reader reader;
                 decltype(haploid_genome::n) n;
@@ -77,16 +49,16 @@ namespace fwdpp
                 reader(buffer, &n);
                 reader(buffer, &nm);
                 if (nm)
-                   {
-                       mutations.resize(nm);
-                       reader(buffer, mutations.data(), nm);
-                   }
+                    {
+                        mutations.resize(nm);
+                        reader(buffer, mutations.data(), nm);
+                    }
                 reader(buffer, &nm);
                 if (nm)
-                   {
-                       smutations.resize(nm);
-                       reader(buffer, smutations.data(), nm);
-                   }
+                    {
+                        smutations.resize(nm);
+                        reader(buffer, smutations.data(), nm);
+                    }
                 return haploid_genome(n, std::move(mutations), std::move(smutations));
             }
         };
