+++ fwdpp/fundamental_types/Makefile.am
@@ -1,0 +1,4 @@
+pkginclude_HEADERS=typedefs.hpp \
+				   haploid_genome.hpp \
+				   mutation_base.hpp
+
