+++ fwdpp/fundamental_types/typedefs.hpp
@@ -1,0 +1,12 @@
+#ifndef FWDPP_FUNDAMENTAL_TYPES_TYPEDEFS_HPP__
+#define FWDPP_FUNDAMENTAL_TYPES_TYPEDEFS_HPP__
+
+#include <cstdint>
+
+namespace fwdpp
+{
+    //! The unsigned integer type is 32 bits
+    using uint_t = std::uint32_t;
+}
+
+#endif
